Move media-related views to own module
